set scalar type to double precision

+++ include/readdy/common/common.h
@@ -35,7 +35,7 @@
 #include "logging.h"
 
 NAMESPACE_BEGIN(readdy)
-using scalar = float;
+using scalar = double;
 using time_step_type = unsigned long;
 using particle_type_type = unsigned short;
